\name{disc-methods}
\docType{methods}
\alias{disc-methods}
\alias{disc,mass,numeric-method}
\title{Methods for Function \code{disc}}
\description{
Discount an evidence structure. For more detail see \code{\link{disc}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mass", y = "numeric")}}{
Discount a mass function.
}
}}
\keyword{methods}
