\name{extPoints-methods}
\docType{methods}
\alias{extPoints-methods}
\alias{extPoints,credal-method}
\title{Methods for Function \code{extPoints}}
\description{
Returns the set of extreme points of a credal set. For more detail see \code{\link{extPoints}}.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "credal")}}{
    Returns the set of extreme points
}
}}

\author{
Alexander Karlsson
}
