\name{focal-methods}
\docType{methods}
\alias{focal-methods}
\alias{focal,mass-method}
\title{Methods for Function \code{focal}}
\description{
Methods for function \code{focal}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mass")}}{
Access function for slot focal 
}
}}

\note{See further \code{\link{focal}}}

\keyword{methods}

