\name{focal<--methods}
\docType{methods}
\alias{focal<--methods}
\alias{focal<-,mass,list-method}

\title{Methods for Function \code{focal<-}}
\description{
Replacement function for focal elements. For more detail see \code{\link{focal<-}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mass")}}{
Replace focal elements
}
}}

\keyword{methods}
