\name{lower-methods}
\docType{methods}
\alias{lower-methods}
\alias{lower,credal,character-method}
\alias{lower,credal,missing-method}
\alias{lower,mass,character-method}
\title{Methods for Function \code{lower}}
\description{
Calculate lower bounds for some states with respect to the evidence structure. For more detail see \code{\link{lower}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "credal", set = "character")}}{
obtain the lower bound for a specific set of states
}


\item{\code{signature(x = "mass", set = "character")}}{
obtain the belief, or lower bounds, for a specific set of states
}
}}


\keyword{methods}

