\name{pign-methods}
\docType{methods}
\alias{pign-methods}
\alias{pign,mass-method}
\title{Methods for Function \code{pign}}
\description{
The pignistic transformation transform a mass function to probability function. For more detail see \code{\link{pign}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mass")}}{
Apply the pignistic transformation on a mass function
}
}}
\keyword{methods}

