\name{relPl-methods}
\docType{methods}
\alias{relPl-methods}
\alias{relPl,mass-method}
\title{Methods for Function \code{relPl}}
\description{
The relative plausability transform transforms a mass function to probability function. For more detail see \code{\link{relPl}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "mass")}}{
Apply the relative plausability transform on a mass function
}
}}
\keyword{methods}
