\name{space-methods}
\docType{methods}
\alias{space-methods}
\alias{space,credal-method}
\alias{space,mass-method}
\title{Methods for Function \code{space}}
\description{
Returns the state space for an evidence structure. For more detail see \code{\link{space}}.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "credal")}}{
Returns the state space for a credal set
}

\item{\code{signature(x = "mass")}}{
Returns the state space for a mass function
}
}}
\keyword{methods}
