\name{space<--methods}
\docType{methods}
\alias{space<--methods}
\alias{space<-,credal,character-method}
\alias{space<-,mass,character-method}


\title{Methods for Function \code{space<-}}
\description{
Replace the state space of an evidence structure. For more details see \code{\link{space}}.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "credal")}}{
Replace state space of a credal set
}

\item{\code{signature(x = "mass")}}{
Replace the state space of a mass function
}
}}
\keyword{methods}
