\name{lower-methods}
\docType{methods}
\alias{lower-methods}
\alias{lower,credal,character-method}
\alias{lower,mass,character-method}
\title{Methods for Function \code{lower}}
\description{
Calculate lower bounds for a vector of sets with respect to the evidence structure. For more detail see \code{\link{lower}}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "credal", sets = "numeric")}}{
obtain lower bounds for a vector of sets
}


\item{\code{signature(x = "mass", sets = "numeric")}}{
obtain the belief, or lower bounds, for a vector of sets
}
}}

\keyword{methods}

